requirements.txt
fastapi==0.135.2
uvicorn[standard]==0.42.0
sqlalchemy==2.0.48
python-dotenv==1.2.2
pyjwt==2.12.1
python-multipart==0.0.22
pydantic==2.12.5
email-validator==2.3.0